Add NSWindow sizing and ordering functions
Tested each manually to ensure they work, except a few that are hard to
visually or log verify, for those they at least run without exceptions.

There is one removed function in here that didn't work: Window_setFrame.
Which gives:
'NSInvalidArgumentException', reason: '-[NSWindow setFrame:]: unrecognized selector...'
when used, as frame is a readonly property on the Window.

I've grouped and ordered them according to how they show up in:
https://developer.apple.com/documentation/appkit/nswindow/frame?language=objc
for lack of a better idea.