Strengthen binary_search return properties.
Prior to this commit, the guarantee is that if the element does not exist, then a uniquely
determined position is returned (which is where the element *would* go)
according to the documentation and tests.

However, if the element exists, the documentation states that there is
no uniquely determined position that gets returned.

1. We can make a stronger statement on the returned value when
there are multiple matches given the existing implementation.

2. We can also make our statement more consistent where both cases of
finding and not finding the desired element results in a uniquely
determined return position.

The binary_search and binary_search_by implementations are basically
equivalent to what `std::lower_bound` does in C++. We can examine the code
for the case where the range `[left, right)` is looking at a range of
equal numbers and see that we set `right = mid` which moves the `right)`
interval towards the left (aka an earlier match). The loop terminates
without `[left` ever moving and with `right)` shifting all the way to
equal `[left`, which is at the first match.

I doubt changing the statement from returning anything within a range to
a specific element would break a lot of existing code for a binary
search implementation, but this change would result in clearer
understanding of the behavior and more consistency within the
documentation.